+++ packages/canvas/src/index.ts
@@ -1,4 +1,5 @@
 export { Canvas } from "./controllers/Canvas";
+export type { CanvasExportHandle } from "./controllers/Canvas";
 export { CanvasThumbnail } from "./controllers/CanvasThumbnail";
 export {
 	exportCanvasToPng,